+++ actionfx-controlsfx/src/main/java/module-info.java
@@ -1,0 +1,5 @@
+module actionfx.actionfx.controlsfx.main {
+    requires javafx.base;
+    requires javafx.controls;
+    requires org.controlsfx.controls;
+}