+++ actionfx-spring-boot/src/integration-test/java/module-info.java
@@ -1,0 +1,15 @@
+module actionfx.actionfx.spring.boot.integrationTest {
+    requires javafx.base;
+    requires javafx.controls;
+    requires javafx.fxml;
+    requires org.hamcrest;
+    requires org.junit.jupiter.api;
+    requires spring.beans;
+    requires spring.boot.test;
+    requires spring.context;
+    requires spring.test;
+    requires actionfx.actionfx.core.integrationTest;
+    requires actionfx.actionfx.core.main;
+    requires actionfx.actionfx.spring.boot.main;
+    requires actionfx.actionfx.testing.main;
+}