+++ actionfx-spring-boot/src/main/java/module-info.java
@@ -1,0 +1,9 @@
+module actionfx.actionfx.spring.boot.main {
+    requires org.slf4j;
+    requires spring.beans;
+    requires spring.context;
+    requires spring.core;
+    requires actionfx.actionfx.core.main;
+
+    exports com.github.actionfx.spring.container;
+}