+++ actionfx-testing/src/main/java/module-info.java
@@ -1,0 +1,9 @@
+module actionfx.actionfx.testing.main {
+    requires javafx.graphics;
+    requires org.junit.jupiter.api;
+    requires org.testfx;
+    requires org.testfx.junit5;
+
+    exports com.github.actionfx.testing.annotation;
+    exports com.github.actionfx.testing.junit5;
+}